Firefox OS packaging and icon.
This is the first pass for a Firefox OS packaging.
(Following the tutorial at https://hacks.mozilla.org/2013/06/building-a-todo-app-for-firefox-os-part-1/)

The app installs and runs in the simulator but the CSS is lacking.
A custom CSS needs to be written.

The icon is based on the public domain SVG in the art/ folder.

The source is http://www.clker.com/clipart-10816.html.